admin requests cruds

+++ app/controllers/admin/EmpRequests.php
@@ -12,9 +12,23 @@ class EmpRequests extends Controller
         if ($username != 'User' && $_SESSION['USER']->status == 'admin') {
 
             if ($_SERVER["REQUEST_METHOD"] == "POST") {
-                $arr["id"] = $_GET["id"];
-                $request->delete($arr["id"]);
-                redirect("admin/emprequests");
+
+                if(isset($_POST["delete"])){
+                    $arr["id"] = $_GET["id"];
+                    $request->delete($arr["id"]);
+                    redirect("admin/emprequests");
+                }
+
+                if(isset($_POST["update"])){
+
+                    $id = $_POST["id"];
+                    unset($_POST["id"]);
+                    unset($_POST["update"]);
+                    $request->update($id, $_POST);
+                    redirect("admin/emprequests");
+
+                }
+
             }
 
             $this->view('admin/emprequests', $data);
